Data.HyperLogLog

Description

See the original paper for details: http://algo.inria.fr/flajolet/Publications/FlFuGaMe07.pdf

HyperLogLog

data HyperLogLog p

Initialize a new counter:

>>> mempty :: HyperLogLog $(3)
HyperLogLog {runHyperLogLog = fromList [0,0,0,0,0,0,0,0]}

Please note how you specify a counter size with the $(n) invocation. Sizes of up to 16 are valid, with 7 being a likely good minimum for decent accuracy.

Let's count a list of unique items and get the latest estimate:

>>> size (foldr insert mempty [1..10] :: HyperLogLog $(4))
Approximate {_confidence = 0.9972, _lo = 2, _estimate = 9, _hi = 17}

Note how insert can be used to add new observations to the approximate counter.

class HasHyperLogLog a p | a -> p where

Methods

hyperLogLog :: Lens' a (HyperLogLog p)

size :: ReifiesConfig p => HyperLogLog p -> Approximate Int64

Approximate size of our set

intersectionSize :: ReifiesConfig p => [HyperLogLog p] -> Approximate Int64

insert :: (ReifiesConfig s, Serial a) => a -> HyperLogLog s -> HyperLogLog s

insertHash :: ReifiesConfig s => Word32 -> HyperLogLog s -> HyperLogLog s

Insert a value that has already been hashed by whatever user defined hash function you want.

cast :: forall p q. (ReifiesConfig p, ReifiesConfig q) => HyperLogLog p -> Maybe (HyperLogLog q)

Config

data Config

Constants required for a bucketing factor b

hll :: Int -> Config

Precalculate constants for a given bucketing factor b

ReifiesConfig

class ReifiesConfig o

Minimal complete definition

reflectConfig

reifyConfig :: Int -> (forall o. ReifiesConfig o => Proxy o -> r) -> r